requirements.txt
--extra-index-url https://download.pytorch.org/whl/cpu

fastapi>=0.95.0
uvicorn>=0.20.0
python-multipart
torch>=1.13.0
torchvision>=0.14.0
pillow>=9.0.0
redis>=4.5.0
rq==1.15.0
